wip: refactoring - avoid duplication between non-reg and navigation tests
