+++ test/e2e/navigation.diagram.test.ts
@@ -13,6 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+import { BpmnDiagramPreparation, BpmnLoadMethod, ImageSnapshotConfigurator, ImageSnapshotThresholdConfig } from './helpers/visu-utils';
 
 // TODO use 'jest-image-snapshot' definition types when the lib will be updated
 // The type lib does not support setting config for ssim (4.2.0 released on july 2020)
@@ -30,39 +31,39 @@ declare global {
   }
 }
 
-import { copyFileSync, loadBpmnContentForUrlQueryParam } from '../helpers/file-helper';
-import debugLogger from 'debug';
+// import { copyFileSync, loadBpmnContentForUrlQueryParam } from '../helpers/file-helper';
+// import debugLogger from 'debug';
 
-const log = debugLogger('test');
+// const log = debugLogger('test');
 const graphContainerId = 'bpmn-viewport';
 
 describe('mouse panning', () => {
-  const defaultImageSnapshotConfig = {
-    diffDirection: 'vertical',
-    dumpDiffToConsole: true, // useful on CI (no need to retrieve the diff image, copy/paste image content from logs)
-    // use SSIM to limit false positive
-    // https://github.com/americanexpress/jest-image-snapshot#recommendations-when-using-ssim-comparison
-    comparisonMethod: 'ssim',
-  };
+  // const defaultImageSnapshotConfig = {
+  //   diffDirection: 'vertical',
+  //   dumpDiffToConsole: true, // useful on CI (no need to retrieve the diff image, copy/paste image content from logs)
+  //   // use SSIM to limit false positive
+  //   // https://github.com/americanexpress/jest-image-snapshot#recommendations-when-using-ssim-comparison
+  //   comparisonMethod: 'ssim',
+  // };
+  //
+  // function getSimplePlatformName(): string {
+  //   const platform = process.platform;
+  //   log(`This platform is ${platform}`);
+  //
+  //   if (platform.startsWith('win')) {
+  //     return 'windows';
+  //   } else if (platform.startsWith('darwin')) {
+  //     return 'macos';
+  //   }
+  //   // we don't support other platform than linux, so hardcode it
+  //   return 'linux';
+  // }
 
-  function getSimplePlatformName(): string {
-    const platform = process.platform;
-    log(`This platform is ${platform}`);
-
-    if (platform.startsWith('win')) {
-      return 'windows';
-    } else if (platform.startsWith('darwin')) {
-      return 'macos';
-    }
-    // we don't support other platform than linux, so hardcode it
-    return 'linux';
-  }
-
-  interface ImageSnapshotThresholdConfig {
-    linux: number;
-    macos: number;
-    windows: number;
-  }
+  // interface ImageSnapshotThresholdConfig {
+  //   linux: number;
+  //   macos: number;
+  //   windows: number;
+  // }
 
   /**
    * Configure threshold by bpmn files. When introducing a new test, please don't add threshold until you get failures when running
@@ -72,31 +73,34 @@ describe('mouse panning', () => {
    */
   const imageSnapshotThresholdConfig = new Map<string, ImageSnapshotThresholdConfig>([]);
 
-  function getImageSnapshotConfig(fileName: string): jest.ImageSnapshotConfig {
-    // minimal threshold to make tests for diagram renders pass on local
-    // macOS: Expected image to match or be a close match to snapshot but was 0.00031509446166699817% different from snapshot
-    let failureThreshold = 0.000004;
+  const imageSnapshotConfigurator = new ImageSnapshotConfigurator(imageSnapshotThresholdConfig);
 
-    const config = imageSnapshotThresholdConfig.get(fileName);
-    if (config) {
-      log(`Building dedicated image snapshot configuration for '${fileName}'`);
-      const simplePlatformName = getSimplePlatformName();
-      log(`Simple platform name: ${simplePlatformName}`);
-      // we know here that we have property names related to the 'simple platform name' so ignoring TS complains.
-      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-      // @ts-ignore
-      failureThreshold = config[simplePlatformName];
-    }
+  // function getImageSnapshotConfig(fileName: string): jest.ImageSnapshotConfig {
+  //   // minimal threshold to make tests for diagram renders pass on local
+  //   // macOS: Expected image to match or be a close match to snapshot but was 0.00031509446166699817% different from snapshot
+  //   let failureThreshold = 0.000004;
+  //
+  //   const config = imageSnapshotThresholdConfig.get(fileName);
+  //   if (config) {
+  //     log(`Building dedicated image snapshot configuration for '${fileName}'`);
+  //     const simplePlatformName = getSimplePlatformName();
+  //     log(`Simple platform name: ${simplePlatformName}`);
+  //     // we know here that we have property names related to the 'simple platform name' so ignoring TS complains.
+  //     // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+  //     // @ts-ignore
+  //     failureThreshold = config[simplePlatformName];
+  //   }
+  //
+  //   log(`ImageSnapshot - using failureThreshold: ${failureThreshold}`);
+  //   return { ...defaultImageSnapshotConfig, failureThreshold: failureThreshold, failureThresholdType: 'percent' };
+  // }
 
-    log(`ImageSnapshot - using failureThreshold: ${failureThreshold}`);
-    return { ...defaultImageSnapshotConfig, failureThreshold: failureThreshold, failureThresholdType: 'percent' };
-  }
+  // enum BpmnLoadMethod {
+  //   QueryParam = 'query param',
+  //   Url = 'url',
+  // }
 
-  enum BpmnLoadMethod {
-    QueryParam = 'query param',
-    Url = 'url',
-  }
-
+  // TODO move comment to utils class
   /**
    * Configure how the BPMN file is loaded by the test page.
    *
@@ -108,32 +112,34 @@ describe('mouse panning', () => {
    */
   const bpmnLoadMethodConfig = new Map<string, BpmnLoadMethod>();
 
-  function getBpmnLoadMethod(fileName: string): BpmnLoadMethod {
-    return bpmnLoadMethodConfig.get(fileName) || BpmnLoadMethod.QueryParam;
-  }
+  // function getBpmnLoadMethod(fileName: string): BpmnLoadMethod {
+  //   return bpmnLoadMethodConfig.get(fileName) || BpmnLoadMethod.QueryParam;
+  // }
 
-  function prepareTestResourcesAndGetPageUrl(fileName: string): string {
-    // to have mouse pointer visible during headless test - &showMousePointer=true
-    let url = 'http://localhost:10002/navigation-diagram.html?fitOnLoad=true';
+  const bpmnDiagramPreparation = new BpmnDiagramPreparation(bpmnLoadMethodConfig, { page: 'navigation-diagram' });
 
-    const bpmnLoadMethod = getBpmnLoadMethod(fileName);
-    log(`Use '${bpmnLoadMethod}' as BPMN Load Method for '${fileName}'`);
-    const relPathToBpmnFile = `../fixtures/bpmn/non-regression/${fileName}.bpmn`;
-    switch (bpmnLoadMethod) {
-      case BpmnLoadMethod.QueryParam:
-        const bpmnContent = loadBpmnContentForUrlQueryParam(relPathToBpmnFile);
-        url += `&bpmn=${bpmnContent}`;
-        break;
-      case BpmnLoadMethod.Url:
-        copyFileSync(relPathToBpmnFile, `../../dist/static/diagrams/`, `${fileName}.bpmn`);
-        url += `&url=./static/diagrams/${fileName}.bpmn`;
-        break;
-    }
-    return url;
-  }
+  // function prepareTestResourcesAndGetPageUrl(fileName: string): string {
+  //   // to have mouse pointer visible during headless test - &showMousePointer=true
+  //   let url = 'http://localhost:10002/navigation-diagram.html?fitOnLoad=true';
+  //
+  //   const bpmnLoadMethod = getBpmnLoadMethod(fileName);
+  //   log(`Use '${bpmnLoadMethod}' as BPMN Load Method for '${fileName}'`);
+  //   const relPathToBpmnFile = `../fixtures/bpmn/non-regression/${fileName}.bpmn`;
+  //   switch (bpmnLoadMethod) {
+  //     case BpmnLoadMethod.QueryParam:
+  //       const bpmnContent = loadBpmnContentForUrlQueryParam(relPathToBpmnFile);
+  //       url += `&bpmn=${bpmnContent}`;
+  //       break;
+  //     case BpmnLoadMethod.Url:
+  //       copyFileSync(relPathToBpmnFile, `../../dist/static/diagrams/`, `${fileName}.bpmn`);
+  //       url += `&url=./static/diagrams/${fileName}.bpmn`;
+  //       break;
+  //   }
+  //   return url;
+  // }
 
   it.each(['gateways'])(`%s`, async (fileName: string) => {
-    const url = prepareTestResourcesAndGetPageUrl(fileName);
+    const url = bpmnDiagramPreparation.prepareTestResourcesAndGetPageUrl(fileName);
 
     const response = await page.goto(url);
     // Uncomment the following in case of http error 400 (probably because of a too large bpmn file)
@@ -184,6 +190,6 @@ describe('mouse panning', () => {
 
     const image = await page.screenshot({ fullPage: true });
 
-    expect(image).toMatchImageSnapshot(getImageSnapshotConfig(fileName));
+    expect(image).toMatchImageSnapshot(imageSnapshotConfigurator.getImageSnapshotConfig(fileName));
   });
 });
